refactor: rename pkg
