+++ integration/cmd/osrm-traffic-updater/osrm_traffic_updater.go
@@ -6,7 +6,7 @@ import (
 	"log"
 	"time"
 
-	proxy "github.com/Telenav/osrm-backend/integration/pkg/gen-trafficproxy"
+	proxy "github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
 	"github.com/Telenav/osrm-backend/integration/pkg/trafficproxyclient"
 )
