+++ integration/cmd/trafficproxy-cli/handle_responses.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"time"
 
-	proxy "github.com/Telenav/osrm-backend/integration/pkg/gen-trafficproxy"
+	proxy "github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
 	"github.com/golang/glog"
 )
