+++ integration/cmd/trafficproxy-cli/streaming_delta_receive.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"time"
 
-	proxy "github.com/Telenav/osrm-backend/integration/pkg/gen-trafficproxy"
+	proxy "github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
 	"github.com/Telenav/osrm-backend/integration/pkg/trafficproxyclient"
 	"github.com/golang/glog"
 )
