+++ integration/pkg/trafficproxyclient/get.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"time"
 
-	proxy "github.com/Telenav/osrm-backend/integration/pkg/gen-trafficproxy"
+	proxy "github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
 	"github.com/golang/glog"
 )
